common : Force quote values in hash generation
Currently, if a value in a hash is simply quoted during the generation
of the final file the quote will be removed and the result will be incorrect.
By force quoting them we ensure that in the final file, the value remain quoted.

(cherry picked from commit 705bf02cbc62a3972be5e3024db09f67381ff128)