upgrade/migrate: avoid race condition
Before, compute evacuation worked like this:
* List servers hosted on a compute node
* Migrate servers to another compute node
* Disable nova-compute service on compute node to avoid new server
  scheduling by Nova.

Now, it works like this:
Until there is a server hosted on the compute node, we:
* List servers hosted on a compute node
* Migrate servers to another compute node

Right after we check there is no server hosted on the compute node, we
disable nova-compute.
It allows the case where a server is scheduling during the evacuation,
and that case can occurs an unlimited number of times, so we should not
have a race condition.