type.yaml.tmpl: Fix nova vncproxy params
The commit 856f1c8ca06108992f1d3c22106f1d26a5b5b93a was not enough to
configure novnc_base_url depending on the architure of controller/compute
nodes.
With this patch we are sure to have the same configuration on controller and
compute and not depending on nova::compute and nova::vncproxy parameters.